scvl-ftl-master/scvl-ftl-master/ftl_app/python_requirements.txt
cryptography
python-gnupg
Pillow
markdown
django 
django-tinymce 
django-filter 
djangorestframework 
djangorestframework-recursive 
django-cors-headers 
rdflib
rdflib-jsonld
coreapi
pyyaml